Duplicate `add_ice_candidate`
