+++ rust/connlib/tunnel/src/control_protocol/client.rs
@@ -13,6 +13,7 @@ use connlib_shared::{
 use domain::base::Rtype;
 use ip_network::IpNetwork;
 use secrecy::Secret;
+use webrtc::ice_transport::ice_candidate::RTCIceCandidate;
 use webrtc::ice_transport::{
     ice_parameters::RTCIceParameters, ice_role::RTCIceRole,
     ice_transport_state::RTCIceTransportState, RTCIceTransport,
@@ -61,6 +62,30 @@ fn set_connection_state_update<CB>(
             }
         })
     }));
+}
+
+impl<CB> Tunnel<CB, client::State>
+where
+    CB: Callbacks + 'static,
+{
+    pub async fn add_ice_candidate(
+        &self,
+        conn_id: GatewayId,
+        ice_candidate: RTCIceCandidate,
+    ) -> Result<()> {
+        tracing::info!(%ice_candidate, %conn_id, "adding new remote candidate");
+
+        let peer_connection = self
+            .peer_connections
+            .lock()
+            .get(&conn_id)
+            .ok_or(Error::ControlProtocolError)?
+            .clone();
+        peer_connection
+            .add_remote_candidate(Some(ice_candidate))
+            .await?;
+        Ok(())
+    }
 }
 
 impl<CB> Tunnel<CB, client::State>
